// src/types/json.ts
export interface ILoanOrSalaryAdvance {
  name: string;
  startDate: string | null;
  endDate: string | null;
  payrollOperationFrequency: string;
  customBambooTableRowId: number;
  value: number;
  currencyCode: string;
}

export interface IOtherDeduction {
  customBambooTableRowId: number;
  name: string;
  startDate: string | null;
  endDate: string | null;
  payrollOperationFrequency: string;
  value: number;
  currencyCode: string;
}